refactor (syntax)

+++ src/sample/spring3/_20_annotation_mvc/_06_BindingTest.java
@@ -261,7 +261,7 @@ public class _06_BindingTest extends AbstractDispatcherServletTest {
 		}
 
 		public String getAsText() {
-			return String.valueOf((Integer) this.getValue());
+			return String.valueOf(this.getValue());
 		}
 
 		public void setAsText(String text) throws IllegalArgumentException {
